+++ spb_curate/curate/api/model_diagnosis.py
@@ -13,7 +13,7 @@ from spb_curate.abstract.api.resource import (
 from spb_curate.abstract.superb_ai_object import SuperbAIObject
 from spb_curate.curate.api import settings
 from spb_curate.curate.api.curate import Job
-from spb_curate.curate.api.enums import JobType
+from spb_curate.curate.api.enums import IouType, JobType, Split
 from spb_curate.curate.model.annotation_types import (
     AnnotationType,
     BoundingBox,
@@ -43,7 +43,8 @@ class Diagnosis(CreateResource, PaginateResource):
         dataset_id: str,
         model_name: str,
         class_list: List[str],
-        metadata: Optional[dict] = None,
+        iou_type: IouType,
+        metadata: Optional[Dict[str, Union[int, float, decimal.Decimal]]] = None,
     ) -> Diagnosis:
         """
         Creates a diagnosis.
@@ -56,8 +57,11 @@ class Diagnosis(CreateResource, PaginateResource):
             The name of the model to diagnose.
         class_list
             The list of class names to diagnose.
+        iou_type
+            The IoU type of the diagnosis.
         metadata
             The metadata associated with the diagnosis.
+            Supported fields: ``beta``, ``target_iou``.
         access_key
             An access key for request authentication.
             If provided, overrides the configuration.
@@ -78,6 +82,7 @@ class Diagnosis(CreateResource, PaginateResource):
         param_metadata = {
             "beta": 1.0,
             "class_list": class_list,
+            "iou_type": iou_type,
             "target_iou": 0.5,
         }
 
@@ -318,6 +323,7 @@ class Diagnosis(CreateResource, PaginateResource):
         access_key: Optional[str] = None,
         team_name: Optional[str] = None,
         predictions: List[Prediction],
+        split: Split = Split.VAL,
         asynchronous: bool = True,
     ) -> Job:
         """
@@ -327,6 +333,8 @@ class Diagnosis(CreateResource, PaginateResource):
         ----------
         predictions
             Newly initialized predictions to add.
+        split
+            The subset data type used for training the model.
         asynchronous
             Whether to immediately return the job after creating it.
             If set to ``False``, the function waits for the job to finish before returning.
@@ -347,6 +355,7 @@ class Diagnosis(CreateResource, PaginateResource):
             dataset_id=self.dataset_id,
             diagnosis_id=self.id,
             predictions=predictions,
+            split=split,
             asynchronous=asynchronous,
         )
 
@@ -661,9 +670,11 @@ class Prediction(SuperbAIObject):
             image_id=image_id,
             prediction_class=prediction_class,
             prediction_value=prediction_value,
-            prediction_type=prediction_value._object_type
-            if isinstance(prediction_value, AnnotationType)
-            else prediction_type,
+            prediction_type=(
+                prediction_value._object_type
+                if isinstance(prediction_value, AnnotationType)
+                else prediction_type
+            ),
             **params,
         )
 
@@ -750,6 +761,7 @@ class Prediction(SuperbAIObject):
         dataset_id: str,
         diagnosis_id: str,
         predictions: List[Prediction],
+        split: Split = Split.VAL,
         asynchronous: bool = True,
     ) -> Job:
         """
@@ -763,6 +775,8 @@ class Prediction(SuperbAIObject):
             The ID of the diagnosis to add the predictions to.
         predictions
             Newly initialized predictions to add.
+        split
+            The subset data type used for training the model.
         asynchronous
             Whether to immediately return the job after creating it.
             If set to ``False``, the function waits for the job to finish before returning.
@@ -798,6 +812,7 @@ class Prediction(SuperbAIObject):
                 "dataset_id": dataset_id,
                 "diagnosis_id": diagnosis_id,
                 "predictions": {"param_id": uploaded_param["id"]},
+                "split": split,
             },
         )
 
@@ -814,7 +829,8 @@ def create_diagnosis(
     dataset_id: str,
     model_name: str,
     class_list: List[str],
-    metadata: Optional[dict] = None,
+    iou_type: IouType,
+    metadata: Optional[Dict[str, Union[int, float, decimal.Decimal]]] = None,
 ) -> Diagnosis:
     """
     Creates a diagnosis.
@@ -827,8 +843,11 @@ def create_diagnosis(
         The name of the model to diagnose.
     class_list
         The list of class names to diagnose.
+    iou_type
+        The iou type of the diagnosis.
     metadata
         The metadata associated with the diagnosis.
+        Supported fields: ``beta``, ``target_iou``.
     access_key
         An access key for request authentication.
         If provided, overrides the configuration.
@@ -851,6 +870,7 @@ def create_diagnosis(
         dataset_id=dataset_id,
         model_name=model_name,
         class_list=class_list,
+        iou_type=iou_type,
         metadata=metadata,
     )