+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added train/val split to the prediction upload (sc-3156)
+- Diagnosis metadata docstring clarification (sc-3156)
+
 ## [1.2.2] - 2024-04-11
 
 ### Fixed